+++ app/src/main/java/com/omelan/cofi/components/StepAddCard.kt
@@ -15,6 +15,8 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.testTag
@@ -40,6 +42,7 @@ fun StepAddCard(
     modifier: Modifier = Modifier,
     stepToEdit: Step? = null,
     save: (Step?) -> Unit,
+    onTypeSelect: () -> Unit = {},
     orderInRecipe: Int,
     isLast: Boolean = false,
     isFirst: Boolean = false,
@@ -58,9 +61,14 @@ fun StepAddCard(
         mutableStateOf((stepToEdit?.value ?: 0).toString())
     }
     val textFieldColors = MaterialTheme.createTextFieldColors()
-//    val focusRequester = remember { FocusRequester() }
+    val nameFocusRequester = remember { FocusRequester() }
     val focusManager = LocalFocusManager.current
-
+    val isExpanded = pickedType != null
+    LaunchedEffect(key1 = isExpanded) {
+        if (isExpanded) {
+            onTypeSelect()
+        }
+    }
     fun saveStep() {
         save(
             Step(
@@ -95,7 +103,9 @@ fun StepAddCard(
                     StepType.values().forEach { stepType ->
                         Chip(
                             value = stringResource(id = stepType.stringRes),
-                            onCheck = { pickedType = stepType },
+                            onCheck = {
+                                pickedType = stepType
+                            },
                             isChecked = pickedType == stepType,
                             modifier = Modifier.testTag(
                                 "step_type_button_${stepType.name.lowercase()}"
@@ -104,7 +114,7 @@ fun StepAddCard(
                     }
                 }
             }
-            if (pickedType != null) {
+            if (isExpanded) {
                 OutlinedTextField(
                     label = { Text(text = stringResource(id = R.string.step_add_name)) },
                     value = stepName,
@@ -117,13 +127,13 @@ fun StepAddCard(
                     keyboardActions = KeyboardActions(
                         onNext = {
                             focusManager.moveFocus(FocusDirection.Down)
-                        }
+                        },
                     ),
                     colors = textFieldColors,
                     modifier = Modifier
                         .testTag("step_name")
                         .padding(Spacing.xSmall)
-//                        .focusRequester(focusRequester)
+                        .focusRequester(nameFocusRequester)
                         .fillMaxWidth(),
                 )
                 OutlinedTextField(
@@ -138,10 +148,15 @@ fun StepAddCard(
                         imeAction = if (pickedType?.isNotWaitStepType() == true) {
                             ImeAction.Next
                         } else {
-                            ImeAction.Done
+                            if (stepName.text.isNotBlank()) {
+                                ImeAction.Done
+                            } else {
+                                ImeAction.Previous
+                            }
                         }
                     ),
                     keyboardActions = KeyboardActions(
+                        onPrevious = { focusManager.moveFocus(FocusDirection.Up) },
                         onNext = {
                             focusManager.moveFocus(FocusDirection.Down)
                         },
@@ -165,9 +180,16 @@ fun StepAddCard(
                         singleLine = true,
                         keyboardOptions = KeyboardOptions(
                             keyboardType = KeyboardType.Number,
-                            imeAction = ImeAction.Done
+                            imeAction = if (stepName.text.isNotBlank()) {
+                                ImeAction.Done
+                            } else {
+                                ImeAction.Previous
+                            },
                         ),
-                        keyboardActions = KeyboardActions(onDone = { saveStep() }),
+                        keyboardActions = KeyboardActions(
+                            onDone = { saveStep() },
+                            onPrevious = { nameFocusRequester.requestFocus() }
+                        ),
                         colors = textFieldColors,
                         modifier = Modifier
                             .testTag("step_value")