+++ app/src/main/java/com/omelan/cofi/components/StepItem.kt
@@ -3,9 +3,7 @@ package com.omelan.cofi.components
 import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.*
 import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -16,6 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
 import com.omelan.cofi.R
@@ -35,6 +34,7 @@ fun StepListItem(
     onClick: ((Step) -> Unit)? = null
 ) {
     val constraintModifier = modifier
+        .heightIn(min = 42.dp)
         .animateContentSize()
         .fillMaxWidth()
         .padding(vertical = Spacing.small)